Move sim queries to infinity_factory, fix empty results bug
