+++ contracts/infinity-factory/src/query.rs
@@ -1,9 +1,11 @@
 use crate::helpers::{generate_instantiate_2_addr, index_range_from_query_options};
-use crate::msg::{NextPairResponse, QueryMsg};
+use crate::msg::{NextPairResponse, QueryMsg, QuotesResponse};
 use crate::state::{INFINITY_GLOBAL, SENDER_COUNTER};
 
-use cosmwasm_std::{to_binary, Addr, Binary, Deps, Env, StdResult};
+use cosmwasm_std::{to_binary, Addr, Binary, Deps, Env, StdError, StdResult, Uint128};
 use infinity_global::load_global_config;
+use infinity_pair::helpers::load_payout_context;
+use infinity_pair::pair::Pair;
 use sg_index_query::QueryOptions;
 
 #[cfg(not(feature = "library"))]
@@ -24,6 +26,14 @@ pub fn query(deps: Deps, env: Env, msg: QueryMsg) -> StdResult<Binary> {
             deps.api.addr_validate(&owner)?,
             query_options.unwrap_or_default(),
         )?),
+        QueryMsg::SimSellToPairQuotes {
+            pair,
+            limit,
+        } => to_binary(&query_sim_sell_to_pair_quotes(deps, pair, limit)?),
+        QueryMsg::SimBuyFromPairQuotes {
+            pair,
+            limit,
+        } => to_binary(&query_sim_buy_from_pair_quotes(deps, pair, limit)?),
     }
 }
 
@@ -72,4 +82,78 @@ pub fn query_pairs_by_owner(
     }
 
     Ok(retval)
+}
+
+pub fn query_sim_sell_to_pair_quotes(
+    deps: Deps,
+    mut pair: Pair,
+    limit: u32,
+) -> StdResult<QuotesResponse> {
+    let infinity_global = INFINITY_GLOBAL.load(deps.storage)?;
+    let payout_context = load_payout_context(
+        deps,
+        &infinity_global,
+        &pair.immutable.collection,
+        &pair.immutable.denom,
+    )
+    .map_err(|_| StdError::generic_err("failed to load payout context".to_string()))?;
+
+    let mut quotes: Vec<Uint128> = vec![];
+
+    let mut idx = 0u32;
+
+    pair.update_sell_to_pair_quote_summary(&payout_context);
+    while idx < limit {
+        if let Some(quote_summary) = &pair.internal.sell_to_pair_quote_summary {
+            quotes.push(quote_summary.seller_amount);
+        } else {
+            break;
+        }
+
+        pair.sim_swap_nft_for_tokens(&payout_context);
+
+        idx += 1;
+    }
+
+    Ok(QuotesResponse {
+        denom: pair.immutable.denom,
+        quotes,
+    })
+}
+
+pub fn query_sim_buy_from_pair_quotes(
+    deps: Deps,
+    mut pair: Pair,
+    limit: u32,
+) -> StdResult<QuotesResponse> {
+    let infinity_global = INFINITY_GLOBAL.load(deps.storage)?;
+    let payout_context = load_payout_context(
+        deps,
+        &infinity_global,
+        &pair.immutable.collection,
+        &pair.immutable.denom,
+    )
+    .map_err(|_| StdError::generic_err("failed to load payout context".to_string()))?;
+
+    let mut quotes: Vec<Uint128> = vec![];
+
+    let mut idx = 0u32;
+
+    pair.update_buy_from_pair_quote_summary(&payout_context);
+    while idx < limit {
+        if let Some(quote_summary) = &pair.internal.buy_from_pair_quote_summary {
+            quotes.push(quote_summary.total());
+        } else {
+            break;
+        }
+
+        pair.sim_swap_tokens_for_nft(&payout_context);
+
+        idx += 1;
+    }
+
+    Ok(QuotesResponse {
+        denom: pair.immutable.denom,
+        quotes,
+    })
 }
