+++ contracts/infinity-router/src/query.rs
@@ -1,4 +1,4 @@
-use crate::msg::{QueryMsg, QuotesResponse};
+use crate::msg::QueryMsg;
 use crate::nfts_for_tokens_iterators::{
     iter::NftsForTokens,
     types::{NftForTokensQuote, NftForTokensSource},
@@ -9,9 +9,7 @@ use crate::tokens_for_nfts_iterators::{
     types::{TokensForNftQuote, TokensForNftSource},
 };
 
-use cosmwasm_std::{to_binary, Addr, Binary, Deps, Env, StdError, StdResult, Uint128};
-use infinity_pair::helpers::load_payout_context;
-use infinity_pair::pair::Pair;
+use cosmwasm_std::{to_binary, Addr, Binary, Deps, Env, StdError, StdResult};
 
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::entry_point;
@@ -47,14 +45,6 @@ pub fn query(deps: Deps, env: Env, msg: QueryMsg) -> StdResult<Binary> {
             limit,
             filter_sources.unwrap_or_default(),
         )?),
-        QueryMsg::SimSellToPairQuotes {
-            pair,
-            limit,
-        } => to_binary(&query_sell_to_pair_quotes(deps, pair, limit)?),
-        QueryMsg::SimBuyFromPairQuotes {
-            pair,
-            limit,
-        } => to_binary(&query_buy_from_pair_quotes(deps, pair, limit)?),
     }
 }
 
@@ -91,74 +81,4 @@ pub fn query_tokens_for_nfts(
     let result = iterator.take(limit as usize).collect::<Vec<TokensForNftQuote>>();
 
     Ok(result)
-}
-
-pub fn query_sell_to_pair_quotes(
-    deps: Deps,
-    mut pair: Pair,
-    limit: u32,
-) -> StdResult<QuotesResponse> {
-    let infinity_global = INFINITY_GLOBAL.load(deps.storage)?;
-    let payout_context = load_payout_context(
-        deps,
-        &infinity_global,
-        &pair.immutable.collection,
-        &pair.immutable.denom,
-    )
-    .map_err(|_| StdError::generic_err("failed to load payout context".to_string()))?;
-
-    let mut quotes: Vec<Uint128> = vec![];
-
-    let mut idx = 0u32;
-    while idx < limit {
-        if let Some(quote_summary) = &pair.internal.sell_to_pair_quote_summary {
-            quotes.push(quote_summary.seller_amount);
-        } else {
-            break;
-        }
-
-        pair.sim_swap_nft_for_tokens(&payout_context);
-
-        idx += 1;
-    }
-
-    Ok(QuotesResponse {
-        denom: pair.immutable.denom,
-        quotes,
-    })
-}
-
-pub fn query_buy_from_pair_quotes(
-    deps: Deps,
-    mut pair: Pair,
-    limit: u32,
-) -> StdResult<QuotesResponse> {
-    let infinity_global = INFINITY_GLOBAL.load(deps.storage)?;
-    let payout_context = load_payout_context(
-        deps,
-        &infinity_global,
-        &pair.immutable.collection,
-        &pair.immutable.denom,
-    )
-    .map_err(|_| StdError::generic_err("failed to load payout context".to_string()))?;
-
-    let mut quotes: Vec<Uint128> = vec![];
-
-    let mut idx = 0u32;
-    while idx < limit {
-        if let Some(quote_summary) = &pair.internal.buy_from_pair_quote_summary {
-            quotes.push(quote_summary.total());
-        } else {
-            break;
-        }
-
-        pair.sim_swap_tokens_for_nft(&payout_context);
-
-        idx += 1;
-    }
-
-    Ok(QuotesResponse {
-        denom: pair.immutable.denom,
-        quotes,
-    })
 }
