+++ unit-tests/src/infinity_factory_tests/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod sim_pair_quotes_factory_tests;
