+++ unit-tests/src/infinity_router_tests/mod.rs
@@ -1,6 +1,4 @@
 #[cfg(test)]
 mod nfts_for_tokens_router_tests;
 #[cfg(test)]
-mod sim_pair_quotes_router_tests;
-#[cfg(test)]
 mod tokens_for_nfts_router_tests;
