+++ unit-tests/src/lib.rs
@@ -1,6 +1,8 @@
 #[cfg(test)]
 mod helpers;
 #[cfg(test)]
+mod infinity_factory_tests;
+#[cfg(test)]
 mod infinity_global_tests;
 #[cfg(test)]
 mod infinity_pair_tests;
